+++ include/GTestSuites.hpp
@@ -1,7 +1,7 @@
 /***************************************************************************
  *              GTestSuites.hpp - Test suite container class               *
  * ----------------------------------------------------------------------- *
- *  copyright (C) 2012-2014 by Jean-Baptiste Cayrou                        *
+ *  copyright (C) 2012-2015 by Jean-Baptiste Cayrou                        *
  * ----------------------------------------------------------------------- *
  *                                                                         *
  *  This program is free software: you can redistribute it and/or modify   *
@@ -155,7 +155,7 @@ const GTestSuite* GTestSuites::operator[](const int& index) const
 inline
 int GTestSuites::size(void) const
 {
-    return m_testsuites.size();
+    return (int)m_testsuites.size();
 }
