+++ inst/mwl/include/GMWLSpectrum.hpp
@@ -1,7 +1,7 @@
 /***************************************************************************
  *          GMWLSpectrum.hpp  -  Multi-wavelength spectrum class           *
  * ----------------------------------------------------------------------- *
- *  copyright (C) 2010-2014 by Juergen Knoedlseder                         *
+ *  copyright (C) 2010-2015 by Juergen Knoedlseder                         *
  * ----------------------------------------------------------------------- *
  *                                                                         *
  *  This program is free software: you can redistribute it and/or modify   *
@@ -119,7 +119,7 @@ std::string GMWLSpectrum::classname(void) const
 inline
 int GMWLSpectrum::size(void) const
 {
-    return m_data.size();
+    return (int)m_data.size();
 }
